Major updates to project "Stone to Steel"
- Officially named the project "Stone to Steel."
- Enhanced world generation for greater complexity and immersion.
- Changed the winter biome to a water biome for a refreshed environment.
- Added a jukebox that plays music with random song selection.
- Added new assets to enrich the game experience.
- Fixed and improved the behavior of GUIElements for better interface performance.
- Implemented a player with 8-directional movement animations.
- Introduced a camera system that can lock to the player or operate independently.
- Recoded and redesigned most menus; some settings menus remain in progress.

This update marks a significant milestone with major improvements to gameplay, visuals, and overall polish!

+++ src/main/java/io/github/anthonyclemens/Math/TwoDimensionMath.java
@@ -10,7 +10,7 @@ public class TwoDimensionMath {
         return boundWidth/2f-rect.getWidth()/2f;
     }
 
-    public static float getMiddleX(int itemWidth, int boundWidth){
+    public static float getMiddleX(float itemWidth, float boundWidth){
         return boundWidth/2f-itemWidth/2f;
     }
 }
